Correct the estimation of the distance to the school
In the previous version, we used an approximation of the distance to the school based on the maximum detection range of the sonar in each set. It was a gross approximation, so we went back to the sonar data, calculated the distances from the sonar to the school using ArcGIS and added this info to the dimensiones file.  So now we correct the formula for Misund's correction of the distortion of sonar-calculated along ring measures.